Adjusted the '.then(data => {})' block in 'index.html' to pass user ingredients
TL;DR Making sure that 'selectedIngredients' is defined and accessible where the response is processed. 

----

To integrate the functionality of highlighting user-selecting ingredients in bold within the detailed ingredients of the best match recipe and additional suggested recipes, I needed to modify my existing JavaScript block that handles the response from my Flask application. This involved adjusting the calls to 'createRecipeElement' to include 'selectedIngredients' as an argument.